requirements.txt
py-cord==2.6.1
pytz==2025.2
DateTime==5.5
pyyaml~=6.0
audioop-lts==0.2.1